# target/classes/swagger/adexchange.yaml
...
swagger: '2.0'
info:
  title: TechM Arena Ad Exchange API
  description: API specification for TechM Arena Ad Exchange
  version: "1.0.0"
host: api.test.com
schemes:
  - http
  - https
basePath: /v1
produces:
  - application/json
paths:
...
  /inventory:
    get:
      description: |
        List of Inventory
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - Publisher
      responses:
        200:
          description: List of Inventory
          schema:
            $ref: '#/definitions/InventoryList'          
    post:
      description: Creates a new Inventory.
      operationId: createInventory
      parameters:
        - name: inventory
          in: body
          description: Inventory to be created
          required: true
          schema:
            $ref: '#/definitions/Inventory'
      responses:
        '200':
          description: created Inventory
          schema:
            $ref: '#/definitions/Inventory'
    put:
      description: Updates the Inventory.
      operationId: updateInventory
      parameters:
        - name: inventory
          in: body
          description: Inventory to be updated
          required: true
          schema:
            $ref: '#/definitions/Inventory'
      responses:
        '200':
          description: updated Inventory
          schema:
            $ref: '#/definitions/Inventory'
  /inventory/{id}:
    get:
      description: Returns Inventory for the given id
      operationId: getInventory
      parameters:
        - name: id
          in: path
          description: ID of Inventory
          required: true
          type: integer
      responses:
        '200':
          description: Inventory response
          schema:
            $ref: '#/definitions/Inventory'
    delete:
      description: deletes a single Inventory based on the ID supplied
      operationId: deleteInventory
      parameters:
        - name: id
          in: path
          description: ID of Inventory to delete
          required: true
          type: integer
      responses:
        '200':
          description: Inventory deleted
  /inventory/search:
    get:
      description: Search Inventory
      operationId: searchInventory
      parameters:
        - name: zoneID
          in: query
          description: zoneID
          required: false
          type: string
        - name: digitalPropertyId
          in: query
          description: digitalPropertyId
          required: false
          type: string
        - name: inventoryName
          in: query
          description: inventoryName
          required: false
          type: string
        - name: description
          in: query
          description: description
          required: false
          type: string
        - name: zoneType
          in: query
          description: zoneType
          required: false
          type: string
        - name: adType
          in: query
          description: adType
          required: false
          type: string
        - name: adTemplateID
          in: query
          description: adTemplateID
          required: false
          type: string
        - name: adWidth
          in: query
          description: adWidth
          required: false
          type: integer
        - name: adHeight
          in: query
          description: adHeight
          required: false
          type: integer
        - name: adInvocationTag
          in: query
          description: adInvocationTag
          required: false
          type: string
        - name: passbackAdTag
          in: query
          description: passbackAdTag
          required: false
          type: string
        - name: floorPrice
          in: query
          description: floorPrice
          required: false
          type: integer
        - name: totalRequests
          in: query
          description: totalRequests
          required: false
          type: integer
        - name: totalImpressions
          in: query
          description: totalImpressions
          required: false
          type: integer
        - name: totalRevenues
          in: query
          description: totalRevenues
          required: false
          type: number
          format: float
        - name: keywords
          in: query
          description: keywords
          required: false
          type: string
        - name: status
          in: query
          description: status
          required: false
          type: integer
        - name: createdTimeStart
          in: query
          description: createdTimeStart
          required: false
          type: string
        - name: createdTimeEnd
          in: query
          description: createdTimeEnd
          required: false
          type: string
        - name: updatedTimeStart
          in: query
          description: updatedTimeStart
          required: false
          type: string
        - name: updatedTimeEnd
          in: query
          description: updatedTimeEnd
          required: false
          type: string
        - name: createdByUser
          in: query
          description: createdByUser
          required: false
          type: string
        - name: updatedByUser
          in: query
          description: updatedByUser
          required: false
          type: string
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of Inventory
          schema:
            $ref: '#/definitions/InventoryList'
  /inventory/search/keyword:
    get:
      description: Search Inventory by keyword
      operationId: searchInventoryByKeyword
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of Inventory
          schema:
            $ref: '#/definitions/InventoryList'            
  /inventorySpace:
    get:
      description: |
        List of InventorySpace
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      tags:
        - Publisher
      responses:
        200:
          description: List of InventorySpace
          schema:
            $ref: '#/definitions/InventorySpaceList'          
    post:
      description: Creates a new InventorySpace.
      operationId: createInventorySpace
      parameters:
        - name: inventorySpace
          in: body
          description: InventorySpace to be created
          required: true
          schema:
            $ref: '#/definitions/InventorySpace'
      responses:
        '200':
          description: created InventorySpace
          schema:
            $ref: '#/definitions/InventorySpace'
    put:
      description: Updates the InventorySpace.
      operationId: updateInventorySpace
      parameters:
        - name: inventorySpace
          in: body
          description: InventorySpace to be updated
          required: true
          schema:
            $ref: '#/definitions/InventorySpace'
      responses:
        '200':
          description: updated InventorySpace
          schema:
            $ref: '#/definitions/InventorySpace'
  /inventorySpace/{id}:
    get:
      description: Returns InventorySpace for the given id
      operationId: getInventorySpace
      parameters:
        - name: id
          in: path
          description: ID of InventorySpace
          required: true
          type: integer
      responses:
        '200':
          description: InventorySpace response
          schema:
            $ref: '#/definitions/InventorySpace'
    delete:
      description: deletes a single InventorySpace based on the ID supplied
      operationId: deleteInventorySpace
      parameters:
        - name: id
          in: path
          description: ID of InventorySpace to delete
          required: true
          type: integer
      responses:
        '200':
          description: InventorySpace deleted
  /inventorySpace/search:
    get:
      description: Search InventorySpace
      operationId: searchInventorySpace
      parameters:
        - name: inventoryId
          in: query
          description: inventoryId
          required: false
          type: string
        - name: estimatedValue
          in: query
          description: estimatedValue
          required: false
          type: number
          format: float
        - name: units
          in: query
          description: units
          required: false
          type: string
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of InventorySpace
          schema:
            $ref: '#/definitions/InventorySpaceList'
  /inventorySpace/search/keyword:
    get:
      description: Search InventorySpace by keyword
      operationId: searchInventorySpaceByKeyword
      parameters:
        - name: pageNo
          in: query
          description: pageNo
          required: false
          type: integer
        - name: pageSize
          in: query
          description: pageSize
          required: false
          type: integer
      responses:
        '200':
          description: List of InventorySpace
          schema:
            $ref: '#/definitions/InventorySpaceList'            
...
definitions:
...
  Inventory:
    type: object
    properties:
      id:
        type: integer
        description: Id
      zoneID:
        type: string
      digitalPropertyId:
        type: string
      inventoryName:
        type: string
      description:
        type: string
      zoneType:
        type: string
      adType:
        type: string
      adTemplateID:
        type: string
      adWidth:
        type: integer
      adHeight:
        type: integer
      adInvocationTag:
        type: string
      passbackAdTag:
        type: string
      floorPrice:
        type: integer
      totalRequests:
        type: integer
      totalImpressions:
        type: integer
      totalRevenues:
        type: number
        format: float
      keywords:
        type: string
      status:
        type: integer
      createdTime:
        type: string
        format: date
      updatedTime:
        type: string
        format: date
      createdByUser:
        type: string
      updatedByUser:
        type: string
  InventoryList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      inventorys:
        type: array
        items:
          $ref: '#/definitions/Inventory'      
  InventorySpace:
    type: object
    properties:
      id:
        type: integer
        description: Id
      inventoryId:
        type: string
      estimatedValue:
        type: number
        format: float
      units:
        type: string
  InventorySpaceList:
    type: object
    properties:
      count:
        type: integer
        format: int64
        description: Total count of objects
      inventorySpaces:
        type: array
        items:
          $ref: '#/definitions/InventorySpace'      
...